upload video to oss

+++ src/addons/model/VideoModel.class.php
@@ -89,15 +89,35 @@ class VideoModel extends Model
                 $this->get_video_part($ffmpegpath, $sourceSavePath . '/' . $video_name, $partSavePath . '/' . $video_name);
                 // $result['video_path']   = $from==2 ? $this->_getSavePath().'/'.$video_name : $this->_getSavePath().'/source/'.$video_name;
 
-                $result['video_path'] = $this->_getSavePath() . '/source/' . $video_name;
-                $result['video_mobile_path'] = $this->_getSavePath() . '/source/' . $video_name;
-                $result['video_part_path'] = $this->_getSavePath() . '/part/' . $video_name;
+                //保存到云存储
+                $config = $this->getConfig();
+                $cloud = new AliOss($config['cloud_attach_bucket'], $config['cloud_attach_admin'], $config['cloud_attach_password'], $config['cloud_attach_api_url']);
+                $_savePtah = $this->_getSavePath();
+                $videoPath = $this->_getSavePath() . '/source/' . $video_name;
+                $videoImagePath = $this->_getSavePath() . '/' . $image_name;
+                $videoPartPath = $this->_getSavePath() . '/part/' . $video_name;
+                $ossVideoPath = $_savePtah . '/source/' . $video_name;
+                $ossVideoImgPath = $_savePtah . '/' . $image_name;
+                $ossVideoPartPath = $_savePtah . '/part/' . $video_name;
+                // 保存到oss
+                if (file_exists($videoPath)) {
+                    $uploadOssResult = $cloud->writeFile($ossVideoPath, fopen($videoPath, 'r'));
+                }
+                if (file_exists($videoImagePath)) {
+                    $uploadOssImgResult = $cloud->writeFile($ossVideoImgPath, fopen($videoImagePath, 'r'));
+                }
+                if (file_exists($videoPartPath)) {
+                    $uploadOssPartResult = $cloud->writeFile($ossVideoPartPath, fopen($videoPartPath, 'r'));
+                }
+                $result['video_path'] = $videoPath;
+                $result['video_mobile_path'] = $videoPath;
+                $result['video_part_path'] = $videoPartPath;
                 $result['size'] = intval($_FILES['video']['size']);
                 $result['name'] = t($_FILES['video']['name']);
                 $result['ctime'] = time();
                 $result['uid'] = intval($_SESSION['mid']);
                 $result['extension'] = $video_ext;
-                $result['image_path'] = $this->_getSavePath() . '/' . $image_name;
+                $result['image_path'] = $videoImagePath;
                 $result['transfer_id'] = $transfer_id;
 
                 if ($image_info = getimagesize($savePath . '/' . $image_name)) {
@@ -189,11 +209,13 @@ class VideoModel extends Model
 
         $ossVideoPath = $_savePtah . '/' . $videoName;
         $ossVideoImgPath = $_savePtah . '/' . $imageName;
-        $ossVideoPartPath = $_savePtah . '/' . $videoName;
         // 保存到oss
-        $uploadOssResult = $cloud->writeFile($ossVideoPath, fopen($videoPath, 'r'));
-        $uploadOssImgResult = $cloud->writeFile($ossVideoImgPath, fopen($videoImagePath, 'r'));
-
+        if (file_exists($videoPath)) {
+            $uploadOssResult = $cloud->writeFile($ossVideoPath, fopen($videoPath, 'r'));
+        }
+        if (file_exists($videoImagePath)) {
+            $uploadOssImgResult = $cloud->writeFile($ossVideoImgPath, fopen($videoImagePath, 'r'));
+        }
         $data = array(
             'uid' => isset($_SESSION['mid']) ? intval($_SESSION['mid']) : 0,
             'ctime' => time(),
